examples/vector_store_example/requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
numpy>=1.20.0
# The agents-hub package should be installed with pip install -e . from the repository root
